mpylib: Update for new qstr encoding scheme in .mpy.
This roughly corresponds to changes done to mpy-tools.py in revision
992a6e1dea of MicroPython repo. Qstr's are now encoded inline with
bytecode which reference them (instead of an array after the bytecode).
Processing also switched to operate on .mpy as a stream (instead of
previous implementation, which operated on some parts as stream, and
on some - as buffer, leading to duplication of some routines).