uasyncio.core: For IORead/IOWrite, store stream obj and cancel IO for it.
When cancelling coroutine which is currently in a stream I/O wait state,
we need to remove corresponding stream object from the poller (or otherwise,
it will remain active in poller, and may trigger later, resuming coroutine
in a point where it doesn't expect I/O).

So, instead of storing just False value in such coroutines, store the stream
object in them (via .pend_throw()), and make cancellation routines call
EventLoop.cancel_io() for it.