ast: parser: Parse list comprehensions by reusing generator expr logic.
Use intermediate AST type (GenComp, for generic comprehension) when parsing
a bare "for" expression operator, and then convert it to either GeneratorExp
or ListComp.

Fairly speaking, it might be possible to use recently introduced .parenform
AST attribute to convert GeneratorExp to another comprehension class depending
on context. But for now, prefer a clean two-level approach, where initially
a generic comprehension is parsed, then it's converted to a concrete
comprehension type. Leave tricks with .parenform for later optimizations, if
any.