uasyncio: Merge StreamReader and StreamWriter into a single class.
This dichotomy is largely a CPython's asyncio idiosyncrasy. The classes
are always used in a pair, and contain about the same data fields, so
it makes sense to merge their definition in a single class.

StreamReader and StreamWriter are made aliases of this new Stream class.
The public API stays compatible, i.e. both reader and writer objects are
used with open_connection() and start_server() calls. (And there was
already a convention that .aclose() method should be called on the writer
object of a pair.)

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>