logging: For loggers with NOTSET level, use parent logger.
For this, add Logger.parent field. For now, just set that to the root
logger for any non-root one. This achieves the behavior that if we
don't specify level for individual loggers, but specify it for the
root logger, then the output will happen as expected (which was not
the case).

This change is based on CPython docs
https://docs.python.org/3/library/logging.html#logging.Logger.setLevel :

"When a logger is created, the level is set to NOTSET (which causes
... delegation to the parent when the logger is a non-root logger)."

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>